pre-processing/requirements.txt
nltk==3.8.1
pyenchant==3.2.2
textstat==0.7.4